git-svn-id: svn://localhost/ardour2/branches/midi@875 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/audio_clock.h
@@ -133,10 +133,10 @@ class AudioClock : public Gtk::HBox
 
 	jack_nframes_t last_when;
 
-	long last_hrs;
-	long last_mins;
-	long last_secs;
-	long last_frames;
+	uint32_t last_hrs;
+	uint32_t last_mins;
+	uint32_t last_secs;
+	uint32_t last_frames;
 	bool last_negative;
 
 	long  ms_last_hrs;
